+++ classes/Helper.cryptolens.php
@@ -2,7 +2,7 @@
 namespace Cryptolens_PHP_Client {
     class Helper extends Cryptolens {
 
-        public static function build_params($token, $product_id, $key = null, $machineid = null, array $additional_flags = null){
+        public function build_params($token, $product_id, $key = null, $machineid = null, array $additional_flags = null){
             $parms = array(
                 "token" => $token,
                 "ProductId" => $product_id,
@@ -14,13 +14,35 @@ namespace Cryptolens_PHP_Client {
                 $parms["Key"] = $key;
             };
             if($machineid != null){
-                $parms["MachineCode"] = $machineid;
+                $parms["MachineCode"] = "";# empty string to prevent error
             }
             if($additional_flags != null){
-                $parms = array_merge($parms, $additional_flags);
+                if(is_array($additional_flags)){
+                    foreach($additional_flags as $key => $value){
+                        if(is_array($value)){
+                            $value = implode(",", $value);
+                        } elseif(is_bool($value)){
+                            if($value == true){
+                                $value = "true";
+                            } else {
+                                $value = "false";
+                            }
+                        } elseif(!is_string($value)){
+                            $value = (string) $value;
+                        }
+                        $parms[$key] = $value;
+                    }
+                    #print_r(var_dump($parms));
+                } else {
+                    echo "Parsed \$additional_flags not as an array!";
+                }
             }
             $postfields = ''; $first = true;
             foreach($parms as $i => $x){
+                if (is_array($x)) {
+                    // detect array an skip
+                    continue; // Überspringe die Kodierung für diesen Durchlauf
+                }
                 if($first) { $first = false; } else { $postfields .= '&';}
 
                 $postfields .= urlencode($i) . "=" . urlencode($x);